+++ src/main/java/com/kimje/chat/chats/repository/ChatRoomRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatRepository extends JpaRepository<Chat, Long> {
+public interface ChatRoomRepository extends JpaRepository<Chat, Long> {
 
 	@Query(
 		value = """
